+++ src/realm/object-store/sync/impl/sync_metadata.hpp
@@ -265,8 +265,7 @@ public:
     /// If the platform supports it, setting `should_encrypt` to `true` and not specifying an encryption key will make
     /// the object store handle generating and persisting an encryption key for the metadata database. Otherwise, an
     /// exception will be thrown.
-    SyncMetadataManager(std::string path, bool should_encrypt,
-                        util::Optional<std::vector<char>> encryption_key = none);
+    SyncMetadataManager(std::string path, bool should_encrypt, OwnedBinaryData encryption_key = {});
 
 private:
     SyncUserMetadataResults get_users(bool marked) const;