+++ test/object-store/sync/sync_test_utils.cpp
@@ -234,9 +234,9 @@ struct FakeLocalClientReset : public TestClientReset {
         m_local_config.force_sync_history = true;
         m_remote_config.force_sync_history = true;
         m_local_config.in_memory = true;
-        m_local_config.encryption_key = std::vector<char>();
+        m_local_config.encryption_key = {};
         m_remote_config.in_memory = true;
-        m_remote_config.encryption_key = std::vector<char>();
+        m_remote_config.encryption_key = {};
     }
 
     void run() override